Refactoring test code
During property testing, the model used is representing each record as a
vector of integer id's. By using the union find algorithm, counts for
the number of various types of datomic entities (records, souls, and
matches right now) can be calculated from the original id tuples. This
means that we can submit data to the resolution process and then check
later on, independent of the process, whether the process was correct.

The refactoring used here was making it so that the outside testing process
could use the same information as the resolution process. That meant the
outside process uses datom-like tuples. This switch was made because it
makes turning off matches during testing much easier.