Wrote tests for removal code
Note that the removal code doesn't actually exist yet. Part of writing
the code for invalidating a match is figuring out how to test/verify
that the code works as expected. I've reworked the tests to save
intermediate versions of the in-memory database that can then be
branched and compared to other versions of the db that have different
transactions committed to them. Pretty weird stuff, but it looks like I
can compare what happens when you add and then remove a match to what
happens when you never remove that match at all, which is pretty strange
stuff, all inside test.check.